\name{Matpower}
\alias{Matpower}
\title{Matrix Power}
\description{
Implementation of efficient algorithm to compute the pth power
of a matrix. 
}
\usage{
Matpower(X, p) 
}
\arguments{
  \item{X}{
numeric square matrix
}
  \item{p}{
integer: nonnegative exponent.
}
}
\value{
a matrix containing the pth power of X.
}



\references{
Golub and Van Loan (1983) Matrix Computations. Algorithm 11.2-1. p. 393.
}

\keyword{math}




