\name{MiscMath-package}
\alias{MiscMath-package}
\alias{MiscMath}
\docType{package}
\title{
Miscellaneous Mathematical Tools and Facts
}
\description{
A random assortment of elementary mathematical formulas and calculators. 
}
\keyword{package}

\examples{
# Find the greatest common divisor of 57, 93 and 117. 
gcd(c(57, 93, 117))
}
